tests: refactor MetricsCheck to shared location
I'm about to use it in connection_limits_test.
